+++ test/unit/parse-query/parse-order.test.ts
@@ -11,19 +11,19 @@ describe("parseOrder", () => {
   });
 
   it.each(["asc", "a-sce_nding", "ascending"])("suports %s", (kind) => {
-    parseOrder(searchParams, kind);
+    parseOrder(searchParams, ":", kind);
 
     expect(searchParams.order).toEqual("ascending");
   });
 
   it.each(["desc", "de-sce_nding", "descending"])("suports %s", (kind) => {
-    parseOrder(searchParams, kind);
+    parseOrder(searchParams, "=", kind);
 
     expect(searchParams.order).toEqual("descending");
   });
 
   it("provides error if invalid value is used for order", () => {
-    parseOrder(searchParams, "foo");
+    parseOrder(searchParams, ":", "foo");
 
     expect(searchParams.order).toBeFalsy();
     expect(searchParams.errors[0]).toEqual({
@@ -34,15 +34,26 @@ describe("parseOrder", () => {
   });
 
   it("provides error if order is already specified", () => {
-    parseOrder(searchParams, "asc");
+    parseOrder(searchParams, ":", "asc");
 
-    parseOrder(searchParams, "desc");
+    parseOrder(searchParams, ":", "desc");
 
     expect(searchParams.order).toEqual("ascending");
     expect(searchParams.errors[0]).toEqual({
       key: "order",
       value: "desc",
       message: `Order option "ascending" already chosen. Ordering by "desc" will be ignored.`,
+    });
+  });
+
+  it("provides error if invalid operator is used", () => {
+    parseOrder(searchParams, ">", "asc");
+
+    expect(searchParams.order).toBeFalsy();
+    expect(searchParams.errors[0]).toEqual({
+      key: "order",
+      value: "asc",
+      message: `Order does not support the ">" operator.`,
     });
   });
 });