+++ test/unit/parse-query/parse-sort.test.ts
@@ -18,7 +18,7 @@ describe("parseSort", () => {
     "id",
     "colors",
   ])("suports %s", (kind) => {
-    parseSort(searchParams, kind);
+    parseSort(searchParams, ":", kind);
 
     expect(searchParams.sort).toEqual(kind);
   });
@@ -26,7 +26,7 @@ describe("parseSort", () => {
   it.each(["results", "steps", "prerequisites", "cards"])(
     "supports %s (as alias)",
     (kind) => {
-      parseSort(searchParams, kind);
+      parseSort(searchParams, ":", kind);
 
       expect(searchParams.sort).toEqual(`number-of-${kind}`);
     }
@@ -35,14 +35,14 @@ describe("parseSort", () => {
   it.each(["color", "ci", "coloridentity", "color-identity"])(
     "supports %s as alias for colors",
     (kind) => {
-      parseSort(searchParams, kind);
+      parseSort(searchParams, "=", kind);
 
       expect(searchParams.sort).toEqual("colors");
     }
   );
 
   it("provides error if invalid value is used for sort", () => {
-    parseSort(searchParams, "foo");
+    parseSort(searchParams, ":", "foo");
 
     expect(searchParams.sort).toBeFalsy();
     expect(searchParams.errors[0]).toEqual({
@@ -53,15 +53,26 @@ describe("parseSort", () => {
   });
 
   it("provides error if sort is already specified", () => {
-    parseSort(searchParams, "results");
+    parseSort(searchParams, ":", "results");
 
-    parseSort(searchParams, "cards");
+    parseSort(searchParams, ":", "cards");
 
     expect(searchParams.sort).toEqual("number-of-results");
     expect(searchParams.errors[0]).toEqual({
       key: "sort",
       value: "cards",
       message: `Sort option "number-of-results" already chosen. Sorting by "cards" will be ignored.`,
+    });
+  });
+
+  it("provides error if invalid operator is used", () => {
+    parseSort(searchParams, ">", "cards");
+
+    expect(searchParams.sort).toBeFalsy();
+    expect(searchParams.errors[0]).toEqual({
+      key: "sort",
+      value: "cards",
+      message: `Sort does not support the ">" operator.`,
     });
   });
 });