+++ src/models/list.ts
@@ -13,7 +13,10 @@ export default class SpellbookList extends Array<string> {
     const list = Object.create(SpellbookList.prototype);
 
     if (items) {
-      const entries = items.split(/\.\s?/).filter((entry) => entry);
+      const entries = items
+        .split(/\.\s?/)
+        .map((entry) => entry.replace(/\r?\n|\r/g, "").trim())
+        .filter((entry) => entry);
 
       list.rawString = items;
       list.push(...entries);